requirements.txt
requests
beautifulsoup4
lxml
pymongo
python-dotenv
python-dateutil
yake
PyMuPDF
flask
